checkout: fix case insensitive unknown file check
Summary:
The Rust checkout wasn't detecting a case insensitive unknown file conflict. For example, on a case insensitive file system you have unknown file "foo" locally and your checkout destination wants to add file "FOO" with different contents. Fix by using a map to both detect the conflict using normalized paths, and mapping the unknown file's case to that of the manifest.

This is the easiest thing I could think of to fix it. I think a better fix is to make the manifest case aware somewhat (e.g. fix the "get_ignore_case" method to actually ignore case). From there we have a few options to fix it more wholistically. The first thing I was going to try was to simply normalize the files in the status() output to that of the source/dest manifest. Then the manifest diff stays case sensitive, but would still detect the conflicts.

Reviewed By: sggutier

Differential Revision: D54037553

fbshipit-source-id: f780ef87bc47f340ebc2f8af171f506bac32028e