ext: kill "disablesymlinks" extension
Summary:
Doesn't seemed to be used anymore, and isn't compatible with Rust checkout. (It already wasn't compatible with experimental.nativecheckout, which is turned on everywhere.)

This extension was added originally in D13964546 to skip symlinks for some Windows/NTFS/fuse situation. We have better support for symlinks on Windows, now, so we double don't need this extension.

Reviewed By: sggutier

Differential Revision: D54037549

fbshipit-source-id: 8873a1fa96c6fe47e9f1c11214db476936f3e640