Hide bottom expander if the hunk is near the end
Summary:
This is a small trick to hide the bottom expander in the comparison if there's no lines left to expand. If a hunk was in the middle of a file, it'd have 4 lines of context. If the last hunk has less than that, then it means there's no more context to load, so it's the end of the file.

This can technically miss a case where there's EXACTLY enough context lines, so there's 0 more lines to load, but we'd still show the button. But that's ok.

Reviewed By: quark-zju

Differential Revision: D54315834

fbshipit-source-id: 387fb60548264631cd69f1e61fcf7a9069375ff4