Allow expanding at the bottom of the comparison view
Summary:
We were missing file content expanders at the bottom of the app.

This is because we don't actually know the number of lines in the file, since it's not part of the diff patch format.

We can work around this by just having a generic "expand lines" button, without saying how many lines.

We don't need to show this for added/deleted files, since they have already had everything expanded.

Technically, this button will appear at the bottom even if the hunk is already at the bottom and there's nothing left to load. I don't like that, but it seems like a reasonable limitation.

I had an idea where we could try to look at the diff patch context lines—if the last hunk doesn't have "enough" trailing context lines, then it's becuase it's at the end and there's nothing left to load. I may do this in a follow up.

Reviewed By: quark-zju

Differential Revision: D54314939

fbshipit-source-id: e935e404091bd4fb90cbd5f522128a4a199113a1