Add unified/split mode selector to comparison view
Summary:
Add a button that lets you change the compariosn view between unified and split views. Or, it can be responsive to window width. So on wide screens it uses split, but on narrow screens it usees unified. This better utilizes your available space.

The default is "responsive", which means the default in most window sizes is "split".

This config is persisted to local storage, so you can set it and keep it the same way.

Reviewed By: quark-zju

Differential Revision: D54317951

fbshipit-source-id: d866446c8edb6bbb50c82a5715cb7d8a357d3092