derived_data_service: set ttl on Deriving nodes
Summary:
`Deriving` nodes are zelos ephemeral nodes created by derivation workers before starting to derive an item from the queue to ensure that no two workers are deriving the same item. If a worker encounters an error during derivation it tries to nack the derivation by deleting the `Deriving` node, but if the nack fails the `Deriving` node remains in zelos until the worker is restarted making it impossible for the other workers to attempt to derive it.

This diff experiments with adding a TTL on the `Deriving` nodes to ensure that they don't block derivation indefinitely. This TTL is periodically extended by the workers as long as the derivation is ongoing to allow time for deriving larger items.

Reviewed By: singhsrb

Differential Revision: D54309951

fbshipit-source-id: 866f55c16684709cb6ab74121ab6b8b39305592e