repo api: Replace `prepare_derived_data` with `derive_bulk`, generalizing it
Summary:
With the introduction of the `BulkDerivation` trait on `derived_data_manager` in the previous diff,
`mononoke_api` doesn't need to depend on and enumerate every single derived data type any more to offer the functionality it wants to offer.

Rename `prepare_derived_data` to `derive_bulk` to allow for more transparent abstractions between the derived data world and the mononoke api world.

Reviewed By: markbt

Differential Revision: D54264050

fbshipit-source-id: d853419b38ab07212b2ccf48facc7288c19cbc6d