force poll after queued operations finish
Summary:
We had a bug with forcing polling after operations run. This behavior is important to make the UI update as things are run.

We had code to force a refresh after every operation. But, runOrQueueOperation would exit immediately when an operation is queued, causing the force refresh to happen right away too (fairly uselessly).

Here, I change the semantics so that runOrQueueOperation resolves only when an operation exits... or if it's determined it will never run because of an earlier error.

I think this was always broken, but made much more apparent by the `configHoldOffRefreshMs` that reduces overfetching during the first 10s of a command running.

Note: it seems there's still some stuff to polish here, I've seen cases where progress of the queued submit is not making it to the client.

Reviewed By: quark-zju

Differential Revision: D54338494

fbshipit-source-id: f0cfbbbf51d61295c92f82d72e0e4efe5dccda4d