serverAPIState: add polling to clean up stale "currentOperation"
Summary:
Polling vs callback again. It seems it could be tricky to guarantee that the
callback notification about process exit works. Let's add a "fallback" polling
to ensure the "currentOperation" won't incorrectly spin forever.

The polling uses a Timer that disables itself when there is no running operation.
So in a "silent" state it does not impose extra costs.

Reviewed By: evangrayk

Differential Revision: D54348074

fbshipit-source-id: 4792556dacc61e8c84fd448ac8f09e213cc29169