timer: declaractive setInterval alternative
Summary:
The "Timer" concept is common in classic GUI programming. But seems to be
missing in the JS world. There is `setInterval` but it is not declaractive.
A "Timer" has an (declaractive) "enabled" attribute to control whether it
should work or not. I think this API is nice, partially because setting
`enabled = true` twice does not register duplicated callbacks.

This diff adds a "Timer" that has the `enabled = boolean` API, and integrate
with GC - if the timer object is GC-ed then the callback will auto stop without
worrying about leaking.

Reviewed By: evangrayk

Differential Revision: D54348076

fbshipit-source-id: 2593c2f6b633940fd03155bf14ff804a97f6b3b5