cmdconfig: preserve config order
Summary:
The config items are intended to be ordered in a section. For example, the
order of `[extensions]` is the extension loading order. The same applies to
`[hooks]` execution order. Extension order is sometimes important and wrong
order caused issues in the past.

This diff removes sorting so we can observe the (sometimes important) order
information from the `config` command.

The sort behavior was added by D37512487, with the idea that it makes the
output "more deterministic". However, the order is *already* deterministic
(and we need to, if we randomize the extension loading order it could cause
subtle bugs), so there is no point to sort.

Reviewed By: zzl0

Differential Revision: D54201004

fbshipit-source-id: 96f28b7bfd31fe1894d76bf07036fdc7895ce933