lfs-server: log health checks to Scuba
Summary:
Logging health checks can be noisy, but *not* logging can make debugging hard, which can me in relation to S294022.

This diff makes both successes and failures logging configurable but disabled by default to reduce risk in the rollout, as well as to keep most tests simple. Sampling rates for both are configurable, too.

Changing the order of middlewares works because of the sequence of inbound vs outbound calls. In a nutshell, the existing sequence (ignoring other middleware) was:

```
  inbound
        |
        v
RequestContext middleware
    inbound() checks identities
        |
        v
Scuba middleware
    inbound() prepares context
        |                                     outbound() logs
        v                                           ^
do the work then go back up the stack  -------------|
```

If RequestContext fails, the request drops out of the inbound stack so the Scuba middleware is never invoked.

Swapping things around works for both success and failure:

Success:

```
  inbound
        |
        v
Scuba middleware
    inbound() prepares context
        |                                     outbound() logs
        v                                           ^
request context middleware                          |
    inbound() checks identities                     |
        |                                           |
        v                                           |
do the work then go back up the stack  -------------|
```

 Failure:

```
  inbound
        |
        v
Scuba middleware
    inbound() prepares context
        |                                     outbound() logs
        v                                           ^
request context middleware                          |
    inbound() does not find identities -------------|
```

Obviously there is some risk in this change, but we should be able to catch any issues with a canary in a single region.

Reviewed By: mitrandir77

Differential Revision: D54222779

fbshipit-source-id: bb7eddab5d153a9367b730b7e30966edf1bf5338