Show submit selected button next to selected commits
Summary:
If you multi-select commits, we show a submit button in the sidebar. But that's kind of out of the way, making submitting a partial selection kind of slow.

Instead, let's add the submit selected button next to the bottommost selected commit when you have multiple selected.

We already had this submit selected in the sidebar, so we can refactor it slightly to be reusable.

While doing this, I also fixed a couple of other issues:
- the sidebar multi-select buttons were kind of badly laid out and didn't wrap well. This is fixed by two lines of CSS.
- the update message when multi-selecting was extra fragile because it was selection order-dependent. We can sort the hashes to be resilient to selection order

Reviewed By: quark-zju

Differential Revision: D54397854

fbshipit-source-id: 3afde8e2750d7648172b8d21e8d41a751fafa9e6