Eden cli: Return correct code (failure) if we couldn't release file handles
Summary:
eden rm now has the ability to try to kill processes holding on to files we want to delete.

If the user rejects the offer and prefers to solve things manually, we're showing "success" even though the rm wasn't successful.

This just returns the right code so we don't display that message.

Differential Revision: D54379250

fbshipit-source-id: 8f75a827f2d2c563aecfe29bb8e56dfdbea950d9