Implement support for git pull parameters
Summary: When performing git pull, the response to the fetch command can include a bunch of headers in a particular order based on the request parameters. This diff adds support for those parameters but doesn't fully implement the partial pull. The changes in this diff should work for both git clone and git pull which will be validated by the existing integration test

Differential Revision: D54254333

fbshipit-source-id: d4c9e1d79679315af71766e225d196135172c622