revsets: support resolving revision numbers in Rust
Summary:
I wasn't planning on doing this, but I realized that Rust returns the wrong commit when a revnum looks like a valid hash prefix.

This will make it easier to turn on Rust checkout in existing tests. Previously I was updating revnum usage in "update" commands to use the "desc(...)" revset, but that isn't ideal since it won't actually use the Rust checkout.

Reviewed By: sggutier

Differential Revision: D54037554

fbshipit-source-id: 8755733f912728916f3c5a959efdb21ddc1b873b