derived data: Implement `from_name` for DerivableType
Summary:
`DerivableType` is the canonical struct that represents derived data types, tied at compile time
with the addition of any new derived data types.

In some places, such as config, we use strings to represent the derived data type.

Implement a conversion from a string into this type so we can limit the surface area of "stringly typed" code.

Reviewed By: markbt

Differential Revision: D54370101

fbshipit-source-id: 91cc60238f77828f54c5b8f0661f74bdcdd3352d