move datapackStore creation to EdenMain
Summary:
# Background
The main goal of this stack is to fold `HgBackingStore` layer into`HgQueuedBackingStore`
## This Step
Move HgDatapackStore to HgQueuedBackingstore

`HgBackingStore` creates `HgDatapackStore`. `HgQueuedBackingStore` can access datapackStore through the `HgBackingStore`. To transfer HgBackingStore layer methods to HgQueuedBackingStore, we need to have access to datapackStore in `HgQueuedBackingStore` directly and not through `HgBackingStore`

## This diff
Create `HgDatapackStore` in EdenMain or Test files and pass it down to `HgQueuedBackingStore` with unique pointer and `HgBackingStore` with a raw pointer.
As the lifetime of `HgQueuedBackingStore` and `HgBackingStore` are same then the raw pointer are safe. When the unique pointer delete the owner of raw pointer also deleted.

Reviewed By: genevievehelsel

Differential Revision: D54287373

fbshipit-source-id: f7936665a6c099ed3511476346e872cdf6fa895d