Move retryThreadPool and serverThreadPool to EdenMain
Summary:
# Background
The main goal of this diff stack is to fold `HgBackingStore` layer into`HgQueuedBackingStore`
## This Step
Move `serverThreadPool_` and 'retryThreadPool` to `HGQueuedBackingStore`
## This diff
Create a new constructor for `HGQueuedBackingStore` which is suitable for unit test only. Then pass `serverThreadPool_` to `HGQueuedBackingStore` constructor from `EdenMain.cpp`. It also pass 'retryServerThreadPool_` from EdenMain or from Tests to both `HgQueuedBackingStore` and `HgBackingStore`

Note: Both `retryThreadPool` and `serverThreadPool` are unique pointer that passed to `HgQueuedBackingStore` however a raw pointer of them pass to `HgBackingStore`. It is safe because the life time of these two backing store are same. Also the last diff of this stack remove the `HgBackingStore` and we will remove all the raw pointers.

Reviewed By: genevievehelsel

Differential Revision: D54327898

fbshipit-source-id: 0dbdff501466798134ade72bf4b533999be47c36