Fix commands always entering "forgot" state 5s after exiting
Summary:
Looks like the logic to check for forgotten commands was slightly inaccurate. We don't set `currentOperation` to `null` when it completes, rather, the last command stays there but has its exitCode and exitTime values set. This is how we render the last operation's progress.

So we should not send this operation forgotted check message if the currentOperation is non-null but has exited.

Reviewed By: quark-zju

Differential Revision: D54596067

fbshipit-source-id: 031c9241a10f534a9e46d568565eafaf4d240300