symlinks: recheck files for symlinks when symlinks are disabled
Summary:
Many users were having issues due to an internal tool creating directory symlinks within EdenFS itself **when having symlinks disabled**, which later caused them a number of issues which were surfaced by `eden doctor`.

The reason for all these issues was that symlink creation on Windows is not atomic (see D35235942). This in turn meant that when manually creating a directory symlink, like the internal tool was doing, a notification was received about *something* being created, which is actually later decided by `PrjfsDispatcherImpl::getOnDiskState`:

https://www.internalfb.com/code/fbsource/[403123cc09bd91e927befd3453ec74fc9d908254]/fbcode/eden/fs/inodes/PrjfsDispatcherImpl.cpp?lines=780-782%2C797-803

Since EdenFS was incorrectly seeing directory symlinks only as directories instead as regular files (which it's supposed to do with symlinks disabled) or as symlinks (which is supposed to do with symlinks enabled), this made EdenFS think there was a directory there which was supposed to be checked. This in turn led to trying to look the contents of the target directory, which led to EdenFS see files that were not supposed to be there even though it thought these were present (see first "Found problem" in failed `eden doctor in Test Plan). Additionally, symlinks themselves had the wrong type due to not rechecking their type, as these were being seen as directories (see second "Found problem"  in failed `eden doctor in Test Plan).

D46860405 was originally sent out to deal with this kind of issue, but it turns out this check should also be done even if symlinks were disabled, which this diff changes.

Reviewed By: kmancini

Differential Revision: D54547002

fbshipit-source-id: 2fc9c2ce2e4a209ba6c2ff06a7ab1c3cc516711b