pager: default "less" to "less -FRQX" in Rust
Summary:
We recently added support for custom pagers to native Rust commands. We missed some behavior in the blasted tweakdefaults extension where we automatically convert a plan "less" pager into "less -FRQX". Fix by adding the conversion in-line in Rust.

We were setting the env var LESS=FRX, but that wasn't taking effect if LESS is already set (for me it is set to "-r").

Reviewed By: sggutier

Differential Revision: D54613074

fbshipit-source-id: 5a2884e30509494b5594c904846ea385baef6336