Enable tag fetching based on tag hashes
Summary:
As part of git protocol, we need a way to fetch tags objects based on their object id hashes. This diff adds support for that. Additionally, since the `tag_hash` column was not indexed, this operation could be quite expensive which is why a new index is added on it.

I will also run AOSC for this table

Differential Revision: D54583989

fbshipit-source-id: 5f00a9932b3734c31c239ed7cda74c56d188bce4