eagercompat: don't enable eagercompat w/ lfs extension
Summary:
The legacy lfs extension is not compatible with an eager store. Let's not enable eagercompat mode if the lfs extension is enabled.

The eagercompat mode allows a client repo to get a working edenapi connection to a remote (legacy) revlog repo in tests. This allows modern client features even if the remote repo is legacy.

Reviewed By: liubov-dmitrieva

Differential Revision: D54608519

fbshipit-source-id: 47d67f8d1bbe77d812bb0532bd9cc1b05a87d4b1