pager: ignore SIGPIPE
Summary:
In general when the pager has exited we handle EPIPE from write() calls, so we don't need to raise an exception on SIGPIPE.

I specifically want this because it fixes a crash when exiting a custom pager handled by Rust (Rust can handle custom pagers since D53838087).

Removing the SIGPIPE handler was previously attempted in D14528603, but reverted due to a Python 2 issue. Let's try it again.

Reviewed By: zzl0

Differential Revision: D54616250

fbshipit-source-id: 9f2852805c763478d7a0feac1f2f9180e1f7f227