reveset: change "-" to use journal instead of undo
Summary:
The "-" revset evaluates to your last non-obsolete working copy parent. Previously it used the undo log to find the previous parent, but the Rust checkout doesn't update the undo log. Switch "-" to instead use the journal, which the Rust checkout does update.

I tweaked a few things in the revset:
- Get rid of max max-old-nonobs-commit-candidates config. It doesn't seem necessary with the journal, plus the journal might have to skip over more entries since you get can journal entries for each commit being rebased.
- If there is no previous parent, return empty set instead of raising an error. This is possible now that we don't limit our search.
- Skip previous states with multiple parents. Not sure if this happens, but you probably don't want to re-enter an unfinished merge state.

Reviewed By: quark-zju

Differential Revision: D54400345

fbshipit-source-id: 7c7b35186401c97d931cce34e7e6a4157004777b