Fix for delta cycles for non-disk packs
Summary:
After investigating a bit more on the workings of the git protocol (and reading the Git C code), I have realized the following w.r.t delta resolution:
- For packs generated during git protocol (pull or clone):
     - We cannot have duplicate objects. Git just errors out in this case
     - We CANNOT have delta cycles if we are sending the full repo, i.e. if the cycle exists within the pack and there is no object external to the pack that can be used to break the cycle
     - We CAN have delta cycles if we are sending partial repo (git pull), i.e. the cycle exists in the pack but can be broken by referring an object outside the pack
- For packs generated during git bundle:
     - We cannot have duplicate objects. Git just errors out in this case
     - We CANNOT have delta cycles if we are bundling the full repo AND if we are bundling partial repo, i.e. delta cycles are NEVER allowed

With this realization, I have made changes in the git protocol code to ensure that the cycle problem doesn't come up in git pull / clone. I have also added integration test to ensure that this behavior is tested.

There are a couple of TODOs that I need to address in light of these changes, but I will do those in a separate diff.

Differential Revision: D54595301

fbshipit-source-id: 14261446b7484a64c825d5dbef225cbceeeb278c