Use HashMap with faster hashing algorithm for packfile writer
Summary: We use HashMap to store the set of objects written to packfile to ensure we do not add duplicate objects to the pack. These collections can grow quite large in which case the performance of the hashing algorithm plays a role in the overall runtime. Since the input to these collections are known to be safe, we do not need a cryptographic hash and can instead use a much faster non-cryptographic hashing algorithm used within the rust compiler

Differential Revision: D54638708

fbshipit-source-id: 5053cf94598be3670ca2339878eeacb2a60eddd0