Optimize tag stream generation during git fetch/pull
Summary:
When fetching the stream of annotated tag objects, we get all of them (unless a few were explicitly requested) in case of git clone. In case of git pull/fetch we are supposed to:
- Fetch all the tags that point to anyone of the commits in the stream of commits included in the packfile
- Fetch all the tags that correspond to the `want oid` requirements received from the client

Turns out computing this set is more expensive then just sending all the available tags and it only adds a minor overhead in terms of size (few KBs or couple MBs at most) which is why this diff updates the tag stream fetching logic to fetch all annotated tags. However this decision is not set in stone, we can always go back to the old version if we figure out that works better.

Differential Revision: D54631092

fbshipit-source-id: e98372cba8d0cc0ad067342e622b99fc44b9d555