Add manifest branch name to remapping state
Summary:
Add the manifest branch name to remapping state so that Jellyfish can always see which branch a given aosp-hg diff is based off of.

Remapping state seems like a convenient place to store this info to me, but please let me know if you think there's a better way to do this.

Reviewed By: mitrandir77

Differential Revision: D54611979

fbshipit-source-id: 904342be5f7dfede91d391a2201a85c43c1d5dbf