commitcloud: don't take repo lock when loading backup state
Summary:
This lock conflicts with long running/stuck commands, preventing "cloud backup" or "cloud sync" from running in a timely manner.

I don't think this locking is necessary for correctness. It appears the only thing the lock was protecting was the filternodes() in BackupState.__init__. The filternodes() is making sure our cached set of backed-up heads only contains nodes present in our local changelog. Since the commits have already been backed up, I don't think it really matters if we filter out a node that is currently being added back to the changelog by another command, or if we don't filter out a node that is being removed from the changelog by another command. Similar filtering happens in other places not protected by the lock, anyway. I think the worst that will happen is we attempt to back up a commit we have already backed up (which is fine).

I moved the locking into BackupState.__init__ behind a config flag so we can turn it back on if needed. There were a couple places loading the BackupState without locking the repo, so if we turn the locking back on they would start locking the repo, but I think that is okay.

Reviewed By: jdelliot

Differential Revision: D54656415

fbshipit-source-id: d04ed79595b435eeccce6774ea60eec1b856349f