Avoid using deltas for generating bundle for draft commits
Summary:
One of the use cases for bundles is to help land HG commits in Git. This involves creating a bundle for the draft commit made by the user. Currently the bundle creation logic can lead to delta cycles for partial repo bundles. This is not a problem for git protocol so we won't be resolving it.

In this diff I am changing the bundle generator to not use deltas to ensure that we don't end up with incorrect bundles. Once we have the full functioning git server running, we can deprecate this bundle workflow entirely. For now, this change doesn't impact performance.

Reviewed By: mitrandir77

Differential Revision: D54679715

fbshipit-source-id: c19f1b23928980c59b2fa897adb6ec57bd83f428