Ensure the client WANTS get computed correctly
Summary: During git protocol, we get a list of `WANTS` and a list of `HAVES` which are the set of objects that the client needs and the set of objects that the client already has (respectively). My assumption was that the client would communicate these parameters only in commit IDs, so we would convert these commits ids to bonsais and use them in the commit graph. However, the client could also send the object ids of the annotated tags and expect the server to pick the commit pointed to by the tag instead of the tag itself. So far, we weren't doing that. This diff fixes it.

Reviewed By: mitrandir77

Differential Revision: D54642351

fbshipit-source-id: 3e744d550499bf956bd961e1a7aa1d07c207a434