Add "Apply" button to shelves to pass --keep
Summary:
Sometimes you don't want to remove a shelved change when applying it, so you can keep it around.

Here I add a second button which applies without removing the shelve via `--keep`.

I also considered using a checkbox at the top of the shelved changes UI. I went back and forth on it, but I think a second button is just as good and a little simpler.

Reviewed By: quark-zju

Differential Revision: D54649799

fbshipit-source-id: 6877c632b16943f594dd5f18e31edf3d036533fe