Rename ExecutionContext to RepositoryContext
Summary:
Start of a stack to change how we deal with shared and unshared data in `Repository`s.

Repositories are possibly shared by multiple connections, each of which has its own cwd, logger, tracker, etc. This is stuff that can't be shared. We had partially extracted this into ExecutionContext.

Now, we'll make this ExecutionContext start containing all the rest of the unsharable data from Repository. This will include cached configs, repo fetch age, and other stuff in the future. We'll pass around the context in most places. The connection / ServerToClientApi will always have a ctx to use.

For some shared stuff, the Repository will need to use "any" ctx that was used to access it. So it will keep the ctx used to create it as a best-effort place to use the logger and tracker. But most calls will have a specific non-shared ctx.

To start, this diff just renamed ExecutionContext, since RepositoryContext will be a more accurate name.

Reviewed By: quark-zju

Differential Revision: D54610051

fbshipit-source-id: 479b2ed4ab15cc49914352cef18e256eced02c8e