Rename Repository.ctx to .initialConnectionContext
Summary:
Previously, we saved `ctx` in Repository. But really, this is like "trackerBestEffort", where it's just a best-effort thing. It's the ctx used when the Repository is first created. But if it's reused, that may not be the same ctx used now. Let's use a longer name to discourage using this and make it more clear what the use for it is.

Later in the stack, we'll pass around ctx from the connection instead.

Reviewed By: quark-zju

Differential Revision: D54610042

fbshipit-source-id: 86c095686b195975b3aea3fd605a636bb3376059