Make runCommand require a ctx
Summary: This ctx should be required, so that we always have some context with with to run. In some cases in reused subscription code, the best we have is the initial connection context we setup. This is what it was already using, so there's no behavior difference here, just making the type signature explicit.

Reviewed By: quark-zju

Differential Revision: D54610039

fbshipit-source-id: eddcc34c5826510878b694b9dcd6187090a1254e