Add <Button>
Summary:
Add a custom <Button> implementation, not depending on the VS Code webview toolkit.

API is very similar, though I changed the "appearance" prop to be two props instead. Instead of:
```
<VSCodeButton appearance="primary">...</VSCodeButton>
<VSCodeButton appearance="secondary">...</VSCodeButton>
<VSCodeButton appearance="icon">...</VSCodeButton>
```
You do:
```
<Button primary>...</Button>
<Button>...</Button>
<Button icon>...</Button>
```

This is similar to how you can use `<Button disabled>`. You can also pass a boolean, so `<Button primary={isSave}>` to have `isSave` control if it's blue or grey. I don't feel strongly about this if it's contentious, but I think it's a little easier to type.

I also don't have the concept of "slots", but rather just have the button contents be display: flex with an appropriate gap. So you just include an Icon as your first child in children and it automatically looks correct. I don't think we have much use for icons in children to not be aligned this way...
Also, it integrates with stylex, so you can overwrite any styles you choose, including alignment of the icons.

Reviewed By: quark-zju

Differential Revision: D54662658

fbshipit-source-id: 7bebefbc30eb43858f8981cfc377f4acda88649a