topo_sort: make node parents generic
Summary: `topo_sort::sort_topological` currently takes a dag represented as `GenericMap<T, Vec<T>>` that maps nodes to a `Vec` of their parents. We only ever iterate on references of a node's parents so we can change `Vec<T>` to any `V` that satisfies `for<'a> &'a V: IntoIterator<Item = &'a T>`.

Reviewed By: mitrandir77

Differential Revision: D54683552

fbshipit-source-id: a2574840d9e19df4d223136e7dc97525fec8b5e2