fold Test HgDatapackStoreTest to HgQueuedBackingStoreTest
Summary:
# Background
The main goal of this diff stack is to fold `HgDatapackStore` layer into`HgQueuedBackingStore`

## This Diff
Before removing `HgDatapackStoreTest` we want to fold its test into`HgQueuedBackingStore`.
To do so, this diff creates `struct HgQueuedBackingStoreTestBase` then all the different structs can inherit from this base struct.
Then we will have `HgQueuedBackingStoreNoFaultInjectorTest` which is used by the previous tests in `HgQueuedBackingStoreTest` and we can create two new structs `HgQueuedBackingStoreWithFaultInjectorTest` and `HgQueuedBackingStoreWithFaultInjectorIgnoreConfigTest` to use for the two new tests which are coming from `HgDatapackStoreTest` to test `getTreeBatch`.

Reviewed By: kmancini

Differential Revision: D54550332

fbshipit-source-id: 74e01b86a2dd5cf23722a9d693eed4aea45420c5