Move runtimeOption and store_ to HgQueuedBackingStore
Summary:
# Background
The main goal of this diff stack is to fold `HgDatapackStore` layer into`HgQueuedBackingStore`

## This Diff
Move `runtimeOptions` and `store_` to `HgQueuedBackingStore`.
Same as the previous diffs, we make them a unique pointer which is owned by `HgQueuedBackingStore` and get a raw pointer from them and pass the raw pointer to `HgDatapackStore`.
**Note:** This raw pointer is safe because the lifetime of `HgDatapackStore` (owner of the raw pointer) is bound to `HgQueudBackingStore` (owner of the unique pointer) then in `HgQueuedBackingStore` deconstruction the memory which is owned by the raw pointer will be deleted too.

Reviewed By: jdelliot

Differential Revision: D54445635

fbshipit-source-id: f9535ee23fce5862cd2f8cecbf2d92aaa93e945d