Move getTree* methods to HgQueuedBackingStore
Summary:
# Background
The main goal of this diff stack is to fold `HgDatapackStore` layer into`HgQueuedBackingStore`

## This Diff
We move a couple of methods getTree* getBlob* and getBlobMetadata* to `HgQueuedBackingStore`.
One of the methods (`getTreeBach`) called by `HgDatapackStoreTest`. In this diff we move the `getTreeBach` into `HgQueuedBackingStore` however we keep a copy of this method in `HgDatapackStore` It help us to keep this test happy for now. In the next diff D54550332, the `HgDatapackStoreTest` will be fold in the `HgQueuedBackingStore` therefore we can remove `HgDatapackStoreTest` in the next diff and remove the `getTreeBatch` method from `HgDatapackStore`

Reviewed By: genevievehelsel

Differential Revision: D54506684

fbshipit-source-id: 88331c311f8f7036d0b4c42ed021ba2a06790097