Extract submodule-related config data to separate struct
Summary:
Every time I have to add/modify fields related to git submodules in the small repo config it's a pain, bc this struct is instantiated in a milllion places.

In this no-op diff I extract both fields to a separate struct which implements `Default`.
In the next diff I can add a new field along with a good default for it without updating a ton of callsites.

Reviewed By: clara-9

Differential Revision: D54591932

fbshipit-source-id: eea0480bb35a6c805ffe26795116ff7b5b25f4b7