Automatically repair eden/dirstate current commit divergence
Summary:
This diff handles edenfs failures by resetting the parents in the status and retrying. That's in a nutshell what `eden doctor` would do.

The new behavior is gated on `experimental.repair-eden-dirstate`, which defaults to `true` but can be reverted to `false` in case of issues.

To test this we also add a new `checkout-pre-set-parents` failpoint to trigger the failure scenario.

Reviewed By: muirdm

Differential Revision: D54590605

fbshipit-source-id: 4bac803527ae176951691d1bfaa8c31c6c28bd85