pyrepo: don't hide eagercompat edenapi client
Summary:
In eagercompat mode, a remote revlog repo is able to act as an edenapi server. There are various places in Python that change behavior based on the presence of a repo.nullableedenapi, so initially I "hid" the edenapi from these spots. However, now I want the edenapi accessible for commmitcloud/pull codepaths, so let's unhide it.

I tweaked the Python edenapi clone path to also require remotenames.selectivepull to be enabled since that is implicitly required for edenapi clone (and a lot of legacy tests don't have selective pull enabled).

Reviewed By: liubov-dmitrieva

Differential Revision: D54722981

fbshipit-source-id: 5b64af9af8b7152d773c543ebd24d63cdd0084b8