Look for buggy successions that would change diff numbers
Summary:
We've seen reports of people having commit messages get mixed up. The best we can trace, this could be attributed to succession detection wrongly persisting commit messages between unrelated commits.

Commit succession tracking is intended to persist unsaved edited commit messages if the underlying commit changes.

If an incorrect succession pairing occurs, it could make commit messages get messed up in a hard to understand way.

The idea in this diff is to check if a detected succession has changed the associated diff number. If so, it means it's not a succession we want to use for commit messages, so it should be skipped.

This also adds tracking so we can observe if this is happening for people.

Note that some mutations legitimately change diff numbers, so this is technically prone to false-positives. I consider this edge case OK—the risk is low to miss a succession, but the danger of incorrectly forwarding the wrong commit message is high.

Reviewed By: quark-zju

Differential Revision: D54784362

fbshipit-source-id: 5133ffd2fbb8af364a695872f60af7406dfecaaf