Be stricter about commit info "mode"
Summary:
Subtle bug. We were storing the commit info "mode" in a dedicated atom used to toggle things. This controls whether the commit info shows the head commit for editing, or the blank commit form for making a new commit.

This state was stored regardless of what commit you were looking at. This allows setting it to "commit" mode, then going to a non-head commit, then going back, and it'll still be in commit mode.

However, this allows potential footgun:
if you don't make sure to check `isHead` when checking `mode === 'commit'`, then you may be getting `"commit"` mode behavior on a non-head commit. We had an "isCommitMode" boolean in CommitInfoView, but some places were still using the commit mode directly.

Let's be more strict, and introduce another atom to ensure the "mode" you get is only "commit" when you're looking at the head commit.

I was hoping this would be the source of the "mixed up commit messages" bug, but I've found no repro using this bug.

Reviewed By: quark-zju

Differential Revision: D54779687

fbshipit-source-id: 9890452b891f0e13f279a46c979388b085e43328