tests: fix small issue when running tests with Buck
Summary:
When running tests in opt mode on Windows and macOS, a `.par` file is created which cannot be opened. A recent change made it more comfortable to avoid creating `.testfailed` when running tests outside of the `scm/tests` directory, but it's possible for this directory to not exist in the first directory when tests are being run in opt mode.

This diff asks if that directory exists before trying to create a `.testfile` there, otherwise it reverts to the old behavior

Reviewed By: zzl0

Differential Revision: D54805832

fbshipit-source-id: b77b21107e211582d2686dfaabcaac5b803c1d29